@@ -4,13 +4,14 @@
   "author": "ghom",
   "description": "Discord bot in TypeScript",
   "main": "dist/index.js",
+  "type": "module",
   "scripts": {
     "build": "gulp build",
     "watch": "gulp watch",
     "start": "npm run build && node dist/index",
-    "format": "prettier --write src",
+    "format": "prettier --write src tests",
     "update": "gulp update",
-    "test": "tsc --noEmit"
+    "test": "tsc --noEmit && npm run build && node tests/index.test.js"
   },
   "prettier": {
     "semi": false,
@@ -27,48 +28,51 @@
   "homepage": "https://github.com/CamilleAbella/bot.ts#readme",
   "dependencies": {
     "axios": "^0.21.1",
-    "boxen": "^5.0.1",
-    "chalk": "^4.1.1",
-    "dayjs": "^1.10.5",
-    "discord.js": "^12.5.3",
+    "boxen": "^6.2.1",
+    "chalk": "^5.0.0",
+    "dayjs": "^1.10.7",
+    "discord.js": "^13.3.1",
     "dotenv": "^10.0.0",
     "emoji-regex": "^9.2.2",
-    "figlet": "^1.5.0",
-    "ghom-eval": "^1.0.3",
+    "figlet": "^1.5.2",
+    "ghom-eval": "^1.1.3",
     "ghom-prettify": "^2.0.1",
-    "knex": "^0.95.6",
-    "prettier": "^2.3.0",
+    "knex": "^0.95.14",
+    "prettier": "^2.5.1",
     "regex-parser": "^2.2.11",
     "sqlite3": "^5.0.2",
-    "tims": "^2.0.5",
-    "yargs-parser": "^20.2.7"
+    "tims": "^2.1.0",
+    "yargs-parser": "^21.0.0",
+    "dirname-filename-esm": "^1.1.1"
   },
   "devDependencies": {
     "@types/fancy-log": "^1.3.1",
-    "@types/figlet": "^1.5.1",
-    "@types/gulp-filter": "^3.0.33",
-    "@types/gulp-rename": "^2.0.0",
-    "@types/node": "^15.12.1",
-    "@types/prettier": "^2.2.3",
+    "@types/figlet": "^1.5.4",
+    "@types/gulp-filter": "^3.0.34",
+    "@types/gulp-rename": "^2.0.1",
+    "@types/node": "^17.0.0",
+    "@types/prettier": "^2.4.2",
     "@types/vinyl-paths": "^0.0.31",
-    "@types/ws": "^7.4.4",
-    "@types/yargs-parser": "^20.2.0",
+    "@types/ws": "^8.2.2",
+    "@types/yargs-parser": "^20.2.1",
     "del": "^6.0.0",
-    "discord-api-types": "^0.18.1",
-    "esbuild": "^0.12.6",
+    "discord-api-types": "^0.25.2",
+    "esbuild": "^0.14.5",
     "fancy-log": "^1.3.3",
     "git-commit-info": "^2.0.0",
     "gulp": "^4.0.2",
-    "gulp-esbuild": "^0.8.1",
+    "gulp-esbuild": "^0.9.2",
     "gulp-filter": "^7.0.0",
     "gulp-rename": "^2.0.0",
-    "make-bot.ts": "^1.7.1",
-    "nodemon": "^2.0.7",
-    "typescript": "^4.3.2",
-    "vinyl-paths": "^3.0.1"
+    "make-bot.ts": "^2.4.1",
+    "nodemon": "^2.0.15",
+    "typescript": "^4.5.4",
+    "vinyl-paths": "^4.0.0",
+    "axios": "^0.24.0",
+    "esm": "^3.2.25"
   },
   "engines": {
-    "node": ">=16",
-    "npm": ">=7.5.2"
+    "node": ">=16.6.0",
+    "npm": ">=7.24.2"
   }
 }
