@@ -1,18 +1,23 @@
-const gulp = require("gulp")
-const esbuild = require("gulp-esbuild")
-const filter = require("gulp-filter")
-const vinyl = require("vinyl-paths")
-const rename = require("gulp-rename")
-const del = require("del")
-const log = require("fancy-log")
-const chalk = require("chalk")
-const git = require("git-commit-info")
-const cp = require("child_process")
-const path = require("path")
-const fs = require("fs")
+import fetch from "axios"
+import gulp from "gulp"
+import esbuild from "gulp-esbuild"
+import filter from "gulp-filter"
+import vinyl from "vinyl-paths"
+import rename from "gulp-rename"
+import del from "del"
+import log from "fancy-log"
+import chalk from "chalk"
+import git from "git-commit-info"
+import cp from "child_process"
+import path from "path"
+import fs from "fs"
 
-function gitLog(cb) {
-  const newVersion = git({ cwd: path.join(process.cwd(), "temp") })
+import { dirname } from "dirname-filename-esm"
+
+const __dirname = dirname(import.meta)
+
+function _gitLog(cb) {
+  const newVersion = git({ cwd: path.join(__dirname, "temp") })
 
   log(
     [
@@ -26,26 +31,56 @@ function gitLog(cb) {
   cb()
 }
 
-function cleanDist() {
+function _cleanDist() {
   return del(["dist/**/*"])
 }
 
-function cleanTemp() {
+function _cleanTemp() {
   return del(["temp"], { force: true })
 }
 
-function downloadTemp(cb) {
+function _checkGulpfile(cb) {
+  fetch(
+    "https://raw.githubusercontent.com/CamilleAbella/bot.ts/master/Gulpfile.js"
+  )
+    .then((res) => res.data)
+    .then(async (remote) => {
+      const local = await fs.promises.readFile(
+        path.join(__dirname, "Gulpfile.js"),
+        "utf8"
+      )
+
+      if (remote !== local) {
+        await fs.promises.writeFile(
+          path.join(__dirname, "Gulpfile.js"),
+          remote,
+          "utf8"
+        )
+
+        log(
+          `${chalk.red("Gulpfile updated!")} Please re-run the ${chalk.cyan(
+            "update"
+          )} command.`
+        )
+
+        process.exit(0)
+      } else cb()
+    })
+    .catch(cb)
+}
+
+function _downloadTemp(cb) {
   cp.exec("git clone https://github.com/CamilleAbella/bot.ts.git temp", cb)
 }
 
-function build() {
+function _build() {
   return gulp
     .src("src/**/*.ts")
     .pipe(
       esbuild({
         sourcemap: "inline",
-        format: "cjs",
-        target: "node12",
+        format: "esm",
+        target: "node16",
         loader: {
           ".ts": "ts",
         },
@@ -54,8 +89,8 @@ function build() {
     .pipe(gulp.dest("dist"))
 }
 
-function watch(cb) {
-  const spawn = cp.spawn("nodemon dist/index", { shell: true })
+function _watch(cb) {
+  const spawn = cp.spawn("nodemon dist/index --delay 1", { shell: true })
 
   spawn.stdout.on("data", (data) => {
     console.log(chalk.white(`${data}`.trim()))
@@ -67,10 +102,10 @@ function watch(cb) {
 
   spawn.on("close", () => cb())
 
-  gulp.watch("src/**/*.ts", { delay: 500 }, gulp.series(cleanDist, build))
+  gulp.watch("src/**/*.ts", { delay: 500 }, gulp.series(_cleanDist, _build))
 }
 
-function copyTemp() {
+function _copyTemp() {
   return gulp
     .src(
       [
@@ -79,21 +114,38 @@ function copyTemp() {
         "temp/src/index.ts",
         "temp/.gitattributes",
         "temp/.gitignore",
-        "temp/Gulpfile.js",
+        "temp/.github/workflows/**/*.native.*",
+        "temp/template.env",
         "temp/tsconfig.json",
+        "temp/tests/**/*.js",
         "!temp/src/app/database.ts",
       ],
       { base: "temp" }
     )
-    .pipe(gulp.dest(process.cwd(), { overwrite: true }))
+    .pipe(gulp.dest(__dirname, { overwrite: true }))
 }
 
-function updateDependencies(cb) {
-  const packageJSON = require("./package.json")
-  const newPackageJSON = require("./temp/package.json")
+function _updateDependencies(cb) {
+  const localPackageJSON = JSON.parse(fs.readFileSync("./package.json", "utf8"))
+  const remotePackageJSON = JSON.parse(
+    fs.readFileSync("./temp/package.json", "utf8")
+  )
+
+  localPackageJSON.main = remotePackageJSON.main
+
+  localPackageJSON.engines = {
+    ...localPackageJSON.engines,
+    ...remotePackageJSON.engines,
+  }
+
+  localPackageJSON.scripts = {
+    ...localPackageJSON.scripts,
+    ...remotePackageJSON.scripts,
+  }
+
   for (const baseKey of ["dependencies", "devDependencies"]) {
-    const dependencies = packageJSON[baseKey]
-    const newDependencies = newPackageJSON[baseKey]
+    const dependencies = localPackageJSON[baseKey]
+    const newDependencies = remotePackageJSON[baseKey]
     for (const key of Object.keys(newDependencies)) {
       if (/^(?:sqlite3|pg|mysql2)$/.test(key)) continue
       if (
@@ -104,8 +156,8 @@ function updateDependencies(cb) {
           `Updated  '${chalk.cyan(key)}' [${
             dependencies[key]
               ? `${chalk.blueBright(dependencies[key])} => ${chalk.blueBright(
-                  newDependencies[key]
-                )}`
+                newDependencies[key]
+              )}`
               : chalk.blueBright(newDependencies[key])
           }]`
         )
@@ -118,15 +170,15 @@ function updateDependencies(cb) {
 
   fs.writeFileSync(
     "./package.json",
-    JSON.stringify(packageJSON, null, 2),
+    JSON.stringify(localPackageJSON, null, 2),
     "utf8"
   )
 
   cp.exec("npm i", cb)
 }
 
-function updateDatabaseFile() {
-  const packageJSON = require("./package.json")
+function _updateDatabaseFile() {
+  const packageJSON = JSON.parse(fs.readFileSync("./package.json", "utf8"))
   const database = ["mysql2", "sqlite3", "pg"].find(
     (name) => name in packageJSON.dependencies
   )
@@ -136,28 +188,36 @@ function updateDatabaseFile() {
     .pipe(gulp.dest("src/app"))
 }
 
-function removeDuplicates() {
+function _removeDuplicates() {
   return gulp
-    .src(["src/**/*.native.ts", "!src/app.native.ts"])
+    .src([
+      "src/**/*.native.ts",
+      "!src/app.native.ts",
+      "temp/.github/workflows/**/*.native.*",
+    ])
     .pipe(
       filter((file) =>
         fs.existsSync(
-          path.join(file.dirname, file.basename.replace("native.ts", "ts"))
+          path.join(
+            file.dirname,
+            file.basename.replace(".native" + file.extname, file.extname)
+          )
         )
       )
     )
     .pipe(vinyl(del))
 }
 
-exports.build = gulp.series(cleanDist, build)
-exports.watch = gulp.series(cleanDist, build, watch)
-exports.update = gulp.series(
-  cleanTemp,
-  downloadTemp,
-  copyTemp,
-  removeDuplicates,
-  updateDependencies,
-  updateDatabaseFile,
-  gitLog,
-  cleanTemp
+export const build = gulp.series(_cleanDist, _build)
+export const watch = gulp.series(_cleanDist, _build, _watch)
+export const update = gulp.series(
+  _checkGulpfile,
+  _cleanTemp,
+  _downloadTemp,
+  _copyTemp,
+  _removeDuplicates,
+  _updateDependencies,
+  _updateDatabaseFile,
+  _gitLog,
+  _cleanTemp
 )
