@@ -1,10 +1,9 @@
-import * as app from "../app"
+import * as app from "../app.js"
 
-import guilds from "../tables/guilds.native"
+import guilds from "../tables/guilds.native.js"
 
-module.exports = new app.Command({
+export default new app.Command({
   name: "prefix",
-  isDefault: true,
   guildOwnerOnly: true,
   channelType: "guild",
   description: "Edit or show the bot prefix",
