@@ -1,12 +1,13 @@
-import * as app from "../app"
+import * as app from "../app.js"
 
-import guilds from "../tables/guilds.native"
+import guilds from "../tables/guilds.native.js"
 
 const listener: app.Listener<"guildDelete"> = {
   event: "guildDelete",
+  description: "Remove guild from db",
   async run(guild) {
     await guilds.query.delete().where("id", guild.id)
   },
 }
 
-module.exports = listener
+export default listener
