@@ -1,22 +1,64 @@
-import * as app from "../app"
+import * as app from "../app.js"
 import yargsParser from "yargs-parser"
+import { filename } from "dirname-filename-esm"
 
-const listener: app.Listener<"message"> = {
-  event: "message",
+const __filename = filename(import.meta)
+
+const listener: app.Listener<"messageCreate"> = {
+  event: "messageCreate",
+  description: "Handle message for commands",
   async run(message) {
-    if (!app.isCommandMessage(message)) return
+    if (!app.isNormalMessage(message)) return
+
+    const prefix = await app.prefix(message.guild ?? undefined)
+
+    if (new RegExp(`^<@!?${message.client.user.id}>$`).test(message.content))
+      return message.channel
+        .send({
+          embeds: [
+            new app.SafeMessageEmbed()
+              .setColor()
+              .setDescription(`My prefix is \`${prefix}\``),
+          ],
+        })
+        .catch()
+
+    message.usedAsDefault = false
+
+    message.send = async function (
+      this: app.NormalMessage,
+      sent: app.SentItem
+    ) {
+      return this.channel.send(sent)
+    }.bind(message)
+
+    message.sendTimeout = async function (
+      this: app.NormalMessage,
+      timeout: number,
+      sent: app.SentItem
+    ) {
+      const m = await this.channel.send(sent)
+      setTimeout(
+        function (this: app.NormalMessage) {
+          if (!this.deleted) this.delete().catch()
+        }.bind(this),
+        timeout
+      )
+      return m
+    }.bind(message)
+
+    message.isFromBotOwner = message.author.id === process.env.BOT_OWNER
 
     app.emitMessage(message.channel, message)
     app.emitMessage(message.author, message)
 
     if (app.isGuildMessage(message)) {
+      message.isFromGuildOwner =
+        message.isFromBotOwner || message.guild.ownerId === message.author.id
+
       app.emitMessage(message.guild, message)
       app.emitMessage(message.member, message)
     }
-
-    message.usedAsDefault = false
-
-    const prefix = await app.prefix(message.guild ?? undefined)
 
     let dynamicContent = message.content
 
@@ -105,13 +147,16 @@ const listener: app.Listener<"message"> = {
       parsedArgs,
       key,
     })
-    if (typeof prepared !== "boolean") return message.channel.send(prepared)
+
+    if (typeof prepared !== "boolean")
+      return message.channel.send({ embeds: [prepared] }).catch()
+
     if (!prepared) return
 
     try {
       await cmd.options.run.bind(cmd)(message)
-    } catch (error) {
-      app.error(error, "handler", true)
+    } catch (error: any) {
+      app.error(error, cmd.filepath ?? __filename, true)
       message.channel
         .send(
           app.code.stringify({
@@ -122,10 +167,10 @@ const listener: app.Listener<"message"> = {
           })
         )
         .catch((error) => {
-          app.error(error, "system")
+          app.error(error, cmd.filepath ?? __filename, true)
         })
     }
   },
 }
 
-module.exports = listener
+export default listener
