@@ -1,7 +1,8 @@
-import * as app from "../app"
+import * as app from "../app.js"
 
 const listener: app.Listener<"messageDelete"> = {
   event: "messageDelete",
+  description: "Remove existing paginator",
   async run(message) {
     const paginator = app.Paginator.getByMessage(message)
 
@@ -9,4 +10,4 @@ const listener: app.Listener<"messageDelete"> = {
   },
 }
 
-module.exports = listener
+export default listener
