@@ -1,19 +1,15 @@
-import * as app from "../app"
+import * as app from "../app.js"
 
 const listener: app.Listener<"messageReactionAdd"> = {
   event: "messageReactionAdd",
+  description: "Handle reaction for paginator",
   async run(reaction, user) {
-    if (!user.bot) {
-      const message = reaction.message
-      const guild = message.guild
-      if (guild) {
-        const paginator = app.Paginator.getByMessage(message)
-        if (paginator) {
-          paginator.handleReaction(reaction, user)
-        }
-      }
-    }
+    if (user.bot) return
+
+    const paginator = app.Paginator.getByMessage(reaction.message)
+
+    if (paginator) return paginator.handleReaction(reaction, user)
   },
 }
 
-module.exports = listener
+export default listener
