@@ -1,17 +1,19 @@
-export * from "./app/pagination"
-export * from "./app/argument"
-export * from "./app/database"
-export * from "./app/listener"
-export * from "./app/command"
-export * from "./app/handler"
-export * from "./app/logger"
-export * from "./app/core"
+export * from "./app/pagination.js"
+export * from "./app/argument.js"
+export * from "./app/database.js"
+export * from "./app/listener.js"
+export * from "./app/command.js"
+export * from "./app/handler.js"
+export * from "./app/logger.js"
+export * from "./app/table.js"
+export * from "./app/core.js"
 
-export * as pagination from "./app/pagination"
-export * as argument from "./app/argument"
-export * as database from "./app/database"
-export * as listener from "./app/listener"
-export * as command from "./app/command"
-export * as handler from "./app/handler"
-export * as logger from "./app/logger"
-export * as core from "./app/core"
+export * as pagination from "./app/pagination.js"
+export * as argument from "./app/argument.js"
+export * as database from "./app/database.js"
+export * as listener from "./app/listener.js"
+export * as command from "./app/command.js"
+export * as handler from "./app/handler.js"
+export * as logger from "./app/logger.js"
+export * as table from "./app/table.js"
+export * as core from "./app/core.js"
