@@ -1,25 +1,25 @@
 import discord from "discord.js"
 import yargsParser from "yargs-parser"
 import regexParser from "regex-parser"
-import emojiRegex from "emoji-regex/es2015/RGI_Emoji"
 
-import * as core from "./core"
-import * as command from "./command"
+import * as core from "./core.js"
+import * as command from "./command.js"
 
 export interface Argument {
   name: string
   description: string
 }
 
-export interface Rest<Message extends command.CommandMessage> extends Argument {
+export interface Rest<Message extends command.NormalMessage> extends Argument {
   required?: core.Scrap<boolean, [message?: Message]>
   default?: core.Scrap<string, [message?: Message]>
   all?: boolean
+  missingErrorMessage?: string | discord.MessageEmbed
 }
 
-export interface Option<Message extends command.CommandMessage>
+export interface Option<Message extends command.NormalMessage>
   extends Argument {
-  aliases?: string[] | string
+  aliases?: string[]
   default?: core.Scrap<string, [message?: Message]>
   required?: core.Scrap<boolean, [message?: Message]>
   castValue?:
@@ -48,40 +48,45 @@ export interface Option<Message extends command.CommandMessage>
     boolean | string,
     [value: any, message?: Message]
   >
-  typeDescription?: core.Scrap<string, [value: string, message?: Message]>
+  castingDescription?: core.Scrap<string, [value: string, message?: Message]>
+  checkingErrorMessage?: string | discord.MessageEmbed
+  castingErrorMessage?: string | discord.MessageEmbed
+  missingErrorMessage?: string | discord.MessageEmbed
 }
 
-export type Positional<Message extends command.CommandMessage> = Omit<
+export type Positional<Message extends command.NormalMessage> = Omit<
   Option<Message>,
   "aliases"
 >
 
-export interface Flag<Message extends command.CommandMessage>
+export interface Flag<Message extends command.NormalMessage>
   extends Pick<Option<Message>, "name" | "aliases" | "description"> {
   flag: string
 }
 
-export function resolveGivenArgument<Message extends command.CommandMessage>(
+export function resolveGivenArgument<Message extends command.NormalMessage>(
   parsedArgs: yargsParser.Arguments,
-  arg: Option<Message> | Flag<Message>
-): { given: boolean; usedName: string; value: any } {
+  arg: Option<any> | Flag<any>
+): {
+  given: boolean
+  nameIsGiven: boolean
+  usedName: string
+  value: any
+} {
   let usedName = arg.name
-  let given = parsedArgs.hasOwnProperty(arg.name)
+  let nameIsGiven = parsedArgs.hasOwnProperty(arg.name)
+  let given =
+    parsedArgs[arg.name] !== undefined && parsedArgs[arg.name] !== null
   let value = parsedArgs[arg.name]
 
   if (!given && arg.aliases) {
-    if (typeof arg.aliases === "string") {
-      usedName = arg.aliases
-      given = parsedArgs.hasOwnProperty(arg.aliases)
-      value = parsedArgs[arg.aliases]
-    } else {
-      for (const alias of arg.aliases) {
-        if (parsedArgs.hasOwnProperty(alias)) {
-          usedName = alias
-          given = true
-          value = parsedArgs[alias]
-          break
-        }
+    for (const alias of arg.aliases) {
+      if (parsedArgs.hasOwnProperty(alias)) {
+        usedName = alias
+        nameIsGiven = true
+        given = true
+        value = parsedArgs[alias]
+        break
       }
     }
   }
@@ -92,11 +97,11 @@ export function resolveGivenArgument<Message extends command.CommandMessage>(
     usedName = arg.flag
   }
 
-  return { given, usedName, value }
+  return { given, usedName, value, nameIsGiven }
 }
 
-export async function checkValue<Message extends command.CommandMessage>(
-  subject: Pick<Option<Message>, "checkValue" | "name">,
+export async function checkValue<Message extends command.NormalMessage>(
+  subject: Pick<Option<any>, "checkValue" | "name">,
   subjectType: "positional" | "argument",
   value: string,
   message: Message
@@ -105,7 +110,7 @@ export async function checkValue<Message extends command.CommandMessage>(
 
   if (Array.isArray(subject.checkValue)) {
     if (subject.checkValue.includes(value)) {
-      return new discord.MessageEmbed()
+      return new core.SafeMessageEmbed()
         .setColor("RED")
         .setAuthor(
           `Bad ${subjectType} pattern "${subject.name}".`,
@@ -124,7 +129,7 @@ export async function checkValue<Message extends command.CommandMessage>(
   )
 
   if (typeof checkResult === "string") {
-    return new discord.MessageEmbed()
+    return new core.SafeMessageEmbed()
       .setColor("RED")
       .setAuthor(
         `Bad ${subjectType} tested "${subject.name}".`,
@@ -135,7 +140,7 @@ export async function checkValue<Message extends command.CommandMessage>(
 
   if (typeof checkResult === "boolean") {
     if (!checkResult) {
-      return new discord.MessageEmbed()
+      return new core.SafeMessageEmbed()
         .setColor("RED")
         .setAuthor(
           `Bad ${subjectType} tested "${subject.name}".`,
@@ -158,7 +163,7 @@ export async function checkValue<Message extends command.CommandMessage>(
   }
 
   if (!checkResult.test(value)) {
-    return new discord.MessageEmbed()
+    return new core.SafeMessageEmbed()
       .setColor("RED")
       .setAuthor(
         `Bad ${subjectType} pattern "${subject.name}".`,
@@ -169,53 +174,61 @@ export async function checkValue<Message extends command.CommandMessage>(
   return true
 }
 
-export async function checkCastedValue<Message extends command.CommandMessage>(
-  subject: Pick<Option<Message>, "checkCastedValue" | "name">,
+export async function checkCastedValue<Message extends command.NormalMessage>(
+  subject: Pick<
+    Option<any>,
+    "checkCastedValue" | "name" | "checkingErrorMessage"
+  >,
   subjectType: "positional" | "argument",
-  value: string,
+  castedValue: any,
   message: Message
 ): Promise<discord.MessageEmbed | true> {
   if (!subject.checkCastedValue) return true
 
   const checkResult: string | boolean = await core.scrap(
     subject.checkCastedValue,
-    value,
+    castedValue,
     message
   )
 
-  if (typeof checkResult === "string") {
-    return new discord.MessageEmbed()
+  const errorEmbed = (errorMessage: string): discord.MessageEmbed => {
+    const embed = new core.SafeMessageEmbed()
       .setColor("RED")
       .setAuthor(
         `Bad ${subjectType} tested "${subject.name}".`,
         message.client.user?.displayAvatarURL()
       )
-      .setDescription(checkResult)
+      .setDescription(errorMessage)
+
+    if (subject.checkingErrorMessage) {
+      if (typeof subject.checkingErrorMessage === "string") {
+        return embed.setDescription(subject.checkingErrorMessage)
+      } else {
+        return subject.checkingErrorMessage
+      }
+    }
+
+    return embed
   }
 
-  if (!checkResult) {
-    return new discord.MessageEmbed()
-      .setColor("RED")
-      .setAuthor(
-        `Bad ${subjectType} tested "${subject.name}".`,
-        message.client.user?.displayAvatarURL()
-      )
-      .setDescription(
-        typeof subject.checkCastedValue === "function"
-          ? core.code.stringify({
-              content: subject.checkCastedValue.toString(),
-              format: true,
-              lang: "js",
-            })
-          : "Please use the `--help` flag for more information."
-      )
-  }
+  if (typeof checkResult === "string") return errorEmbed(checkResult)
+
+  if (!checkResult)
+    return errorEmbed(
+      typeof subject.checkCastedValue === "function"
+        ? core.code.stringify({
+            content: subject.checkCastedValue.toString(),
+            format: true,
+            lang: "js",
+          })
+        : "Please use the `--help` flag for more information."
+    )
 
   return true
 }
 
-export async function castValue<Message extends command.CommandMessage>(
-  subject: Pick<Option<Message>, "castValue" | "name">,
+export async function castValue<Message extends command.NormalMessage>(
+  subject: Pick<Option<any>, "castValue" | "name" | "castingErrorMessage">,
   subjectType: "positional" | "argument",
   baseValue: string | undefined,
   message: Message,
@@ -267,7 +280,20 @@ export async function castValue<Message extends command.CommandMessage>(
             const channel = message.client.channels.cache.get(id)
             if (channel) setValue(channel)
             else throw new Error("Unknown channel!")
-          } else throw new Error("Invalid channel value!")
+          } else {
+            const search = (channel: discord.AnyChannel) => {
+              return (
+                "name" in channel && // @ts-ignore
+                channel.name.toLowerCase().includes(baseValue.toLowerCase())
+              )
+            }
+            let channel: discord.AnyChannel | undefined
+            if (command.isGuildMessage(message))
+              channel = message.guild.channels.cache.find(search)
+            channel ??= message.client.channels.cache.find(search)
+            if (channel) setValue(channel)
+            else throw new Error("Channel not found!")
+          }
         } else throw empty
         break
       case "member":
@@ -276,10 +302,28 @@ export async function castValue<Message extends command.CommandMessage>(
             const match = /^(?:<@!?(\d+)>|(\d+))$/.exec(baseValue)
             if (match) {
               const id = match[1] ?? match[2]
-              const member = message.guild.members.cache.get(id)
+              const member = await message.guild.members.fetch({
+                user: id,
+                force: false,
+                cache: false,
+              })
               if (member) setValue(member)
               else throw new Error("Unknown member!")
-            } else throw new Error("Invalid member value!")
+            } else {
+              const members = await message.guild.members.fetch()
+              const member = members.find((member) => {
+                return (
+                  member.displayName
+                    .toLowerCase()
+                    .includes(baseValue.toLowerCase()) ||
+                  member.user.username
+                    .toLowerCase()
+                    .includes(baseValue.toLowerCase())
+                )
+              })
+              if (member) setValue(member)
+              else throw new Error("Member not found!")
+            }
           } else
             throw new Error(
               'The "GuildMember" casting is only available in a guild!'
@@ -297,7 +341,12 @@ export async function castValue<Message extends command.CommandMessage>(
             const channel = message.client.channels.cache.get(channelID)
             if (channel) {
               if (channel.isText()) {
-                setValue(await channel.messages.fetch(messageID, false))
+                setValue(
+                  await channel.messages.fetch(messageID, {
+                    force: false,
+                    cache: false,
+                  })
+                )
               } else throw new Error("Invalid channel type!")
             } else throw new Error("Unknown channel!")
           } else throw new Error("Invalid message link!")
@@ -308,10 +357,21 @@ export async function castValue<Message extends command.CommandMessage>(
           const match = /^(?:<@!?(\d+)>|(\d+))$/.exec(baseValue)
           if (match) {
             const id = match[1] ?? match[2]
-            const user = await message.client.users.fetch(id, false)
+            const user = await message.client.users.fetch(id, {
+              force: false,
+              cache: false,
+            })
             if (user) setValue(user)
             else throw new Error("Unknown user!")
-          } else throw new Error("Invalid user value!")
+          } else {
+            const user = message.client.users.cache.find((user) => {
+              return user.username
+                .toLowerCase()
+                .includes(baseValue.toLowerCase())
+            })
+            if (user) setValue(user)
+            else throw new Error("User not found!")
+          }
         } else throw empty
         break
       case "role":
@@ -320,10 +380,20 @@ export async function castValue<Message extends command.CommandMessage>(
             const match = /^(?:<@&?(\d+)>|(\d+))$/.exec(baseValue)
             if (match) {
               const id = match[1] ?? match[2]
-              const role = message.guild.roles.cache.get(id)
+              const role = await message.guild.roles.fetch(id)
               if (role) setValue(role)
               else throw new Error("Unknown role!")
-            } else throw new Error("Invalid role value!")
+            } else {
+              const roles = await message.guild.roles.fetch(undefined, {
+                cache: false,
+                force: false,
+              })
+              const role = roles.find((role) => {
+                return role.name.toLowerCase().includes(baseValue.toLowerCase())
+              })
+              if (role) setValue(role)
+              else throw new Error("Role not found!")
+            }
           } else
             throw new Error(
               'The "GuildRole" casting is only available in a guild!'
@@ -339,7 +409,7 @@ export async function castValue<Message extends command.CommandMessage>(
             if (emote) setValue(emote)
             else throw new Error("Unknown emote!")
           } else {
-            const emojiMatch = emojiRegex().exec(baseValue)
+            const emojiMatch = core.emojiRegex.exec(baseValue)
             if (emojiMatch) setValue(emojiMatch[0])
             else throw new Error("Invalid emote value!")
           }
@@ -348,7 +418,7 @@ export async function castValue<Message extends command.CommandMessage>(
       case "invite":
         if (baseValue) {
           if (command.isGuildMessage(message)) {
-            const invites = await message.guild.fetchInvites()
+            const invites = await message.guild.invites.fetch()
             const invite = invites.find(
               (invite) => invite.code === baseValue || invite.url === baseValue
             )
@@ -370,8 +440,29 @@ export async function castValue<Message extends command.CommandMessage>(
   try {
     await cast()
     return true
-  } catch (error) {
-    return new discord.MessageEmbed()
+  } catch (error: any) {
+    const errorCode = core.code.stringify({
+      content: `${error.name}: ${error.message}`,
+      lang: "js",
+    })
+
+    if (subject.castingErrorMessage) {
+      if (typeof subject.castingErrorMessage === "string") {
+        return new core.SafeMessageEmbed()
+          .setColor("RED")
+          .setAuthor(
+            `Bad ${subjectType} type "${subject.name}".`,
+            message.client.user?.displayAvatarURL()
+          )
+          .setDescription(
+            subject.castingErrorMessage.replace(/@error/g, errorCode)
+          )
+      } else {
+        return subject.castingErrorMessage
+      }
+    }
+
+    return new core.SafeMessageEmbed()
       .setColor("RED")
       .setAuthor(
         `Bad ${subjectType} type "${subject.name}".`,
@@ -382,18 +473,13 @@ export async function castValue<Message extends command.CommandMessage>(
           typeof subject.castValue === "function"
             ? "{*custom type*}"
             : "`" + subject.castValue + "`"
-        }\n${core.code.stringify({
-          content: `Error: ${error.message}`,
-          lang: "js",
-        })}`
+        }\n${errorCode}`
       )
   }
 }
 
-export function getTypeDescriptionOf<Message extends command.CommandMessage>(
-  arg: Option<Message>
-) {
-  if (arg.typeDescription) return arg.typeDescription
+export function getCastingDescriptionOf(arg: Option<any>) {
+  if (arg.castingDescription) return arg.castingDescription
   if (!arg.castValue) return "string"
   if (typeof arg.castValue === "string") {
     if (arg.castValue === "array") return "Array<string>"
@@ -402,7 +488,7 @@ export function getTypeDescriptionOf<Message extends command.CommandMessage>(
   return "any"
 }
 
-export function isFlag<Message extends command.CommandMessage>(
+export function isFlag<Message extends command.NormalMessage>(
   arg: Option<Message>
 ): arg is Flag<Message> {
   return arg.hasOwnProperty("flag")
