@@ -5,17 +5,24 @@ import tims from "tims"
 import path from "path"
 import yargsParser from "yargs-parser"
 
-import * as core from "./core"
-import * as logger from "./logger"
-import * as handler from "./handler"
-import * as argument from "./argument"
+import * as core from "./core.js"
+import * as logger from "./logger.js"
+import * as handler from "./handler.js"
+import * as argument from "./argument.js"
+
+import { filename } from "dirname-filename-esm"
+
+const __filename = filename(import.meta)
 
 export const commandHandler = new handler.Handler(
   process.env.BOT_COMMANDS_PATH ?? path.join(process.cwd(), "dist", "commands")
 )
 
-commandHandler.on("load", (filepath) => {
-  return commands.add(require(filepath))
+commandHandler.on("load", async (filepath) => {
+  const file = await import("file://" + filepath)
+  if (filepath.endsWith(".native.js")) file.default.options.native = true
+  file.default.filepath = filepath
+  return commands.add(file.default)
 })
 
 export let defaultCommand: Command<any> | null = null
@@ -26,15 +33,11 @@ export const commands = new (class CommandCollection extends discord.Collection<
 > {
   public resolve(key: string): Command<keyof CommandMessageType> | undefined {
     for (const [name, command] of this) {
-      if (key === name) {
+      if (
+        key === name ||
+        command.options.aliases?.some((alias) => key === alias)
+      )
         return command
-      } else {
-        const aliases = command.options.aliases ?? []
-        const resolvedAliases = Array.isArray(aliases) ? aliases : [aliases]
-        if (resolvedAliases.some((alias) => key === alias)) {
-          return command
-        }
-      }
     }
   }
 
@@ -44,22 +47,32 @@ export const commands = new (class CommandCollection extends discord.Collection<
   }
 })()
 
-export type CommandMessage = discord.Message & {
+export type SentItem = string | discord.MessagePayload | discord.MessageOptions
+
+export type NormalMessage = discord.Message & {
   args: { [name: string]: any } & any[]
   triggerCoolDown: () => void
+  send: (this: NormalMessage, item: SentItem) => Promise<discord.Message>
+  sendTimeout: (
+    this: NormalMessage,
+    timeout: number,
+    item: SentItem
+  ) => Promise<discord.Message>
   usedAsDefault: boolean
+  isFromBotOwner: boolean
+  isFromGuildOwner: boolean
   usedPrefix: string
   client: core.FullClient
   rest: string
 }
 
-export type GuildMessage = CommandMessage & {
+export type GuildMessage = NormalMessage & {
   channel: discord.TextChannel & discord.GuildChannel
   guild: discord.Guild
   member: discord.GuildMember
 }
 
-export type DirectMessage = CommandMessage & {
+export type DirectMessage = NormalMessage & {
   channel: discord.DMChannel
 }
 
@@ -68,16 +81,20 @@ export interface CoolDown {
   trigger: boolean
 }
 
-export type Middleware<
-  Type extends keyof CommandMessageType = keyof CommandMessageType
-> = (
-  message: CommandMessageType[Type]
-) => Promise<boolean | string> | boolean | string
+export interface MiddlewareResult {
+  result: boolean | string
+  data: any
+}
+
+export type Middleware<Type extends keyof CommandMessageType> = (
+  message: CommandMessageType[Type],
+  data: any
+) => Promise<MiddlewareResult> | MiddlewareResult
 
 export interface CommandMessageType {
   guild: GuildMessage
   dm: DirectMessage
-  all: CommandMessage
+  all: NormalMessage
 }
 
 export interface CommandOptions<Type extends keyof CommandMessageType> {
@@ -100,7 +117,7 @@ export interface CommandOptions<Type extends keyof CommandMessageType> {
    * Use this command as slash command
    */
   isSlash?: boolean
-  aliases?: string[] | string
+  aliases?: string[]
   /**
    * Cool down of command (in ms)
    */
@@ -164,9 +181,16 @@ export interface CommandOptions<Type extends keyof CommandMessageType> {
    * @deprecated
    */
   parent?: Command<keyof CommandMessageType>
+  /**
+   * This property is automatically setup on bot running.
+   * @deprecated
+   */
+  native?: boolean
 }
 
 export class Command<Type extends keyof CommandMessageType = "all"> {
+  filepath?: string
+
   constructor(public options: CommandOptions<Type>) {}
 }
 
@@ -186,7 +210,7 @@ export function validateCommand<
         )} command wants to be a default command but the ${chalk.blueBright(
           defaultCommand.options.name
         )} command is already the default command`,
-        "handler"
+        command.filepath ?? __filename
       )
     else defaultCommand = command
   }
@@ -215,12 +239,13 @@ export function validateCommand<
         `you forgot using ${chalk.greenBright(
           "message.triggerCoolDown()"
         )} in the ${chalk.blueBright(command.options.name)} command.`,
-        "handler"
+        "command:validateCommand"
       )
 
   logger.log(
-    `loaded command ${chalk.blueBright(commandBreadcrumb(command))}`,
-    "handler"
+    `loaded command ${chalk.blueBright(commandBreadcrumb(command))}${
+      command.options.native ? ` ${chalk.green("native")}` : ""
+    } ${chalk.grey(command.options.description)}`
   )
 
   if (command.options.subs)
@@ -234,6 +259,7 @@ export function commandBreadcrumb<Type extends keyof CommandMessageType>(
 ): string {
   return commandParents(command)
     .map((cmd) => cmd.options.name)
+    .reverse()
     .join(separator)
 }
 
@@ -241,7 +267,7 @@ export function commandParents<Type extends keyof CommandMessageType>(
   command: Command<Type>
 ): Command<any>[] {
   return command.options.parent
-    ? [command, ...commandParents(command.options.parent)].reverse()
+    ? [command, ...commandParents(command.options.parent)]
     : [command]
 }
 
@@ -279,7 +305,7 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
           trigger: false,
         })
       } else {
-        return new discord.MessageEmbed()
+        return new core.SafeMessageEmbed()
           .setColor("RED")
           .setAuthor(
             `Please wait ${Math.ceil(
@@ -293,7 +319,7 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
     message.triggerCoolDown = () => {
       logger.warn(
         `You must setup the cooldown of the "${cmd.options.name}" command before using the "triggerCoolDown" method`,
-        "system"
+        "command:prepareCommand"
       )
     }
   }
@@ -302,7 +328,7 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
 
   if (isGuildMessage(message)) {
     if (channelType === "dm")
-      return new discord.MessageEmbed()
+      return new core.SafeMessageEmbed()
         .setColor("RED")
         .setAuthor(
           "This command must be used in DM.",
@@ -311,10 +337,10 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
 
     if (core.scrap(cmd.options.guildOwnerOnly, message))
       if (
-        message.guild.owner !== message.member &&
+        message.guild.ownerId !== message.member.id &&
         process.env.BOT_OWNER !== message.member.id
       )
-        return new discord.MessageEmbed()
+        return new core.SafeMessageEmbed()
           .setColor("RED")
           .setAuthor(
             "You must be the guild owner.",
@@ -328,13 +354,8 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
       )
 
       for (const permission of botPermissions)
-        if (
-          !message.guild.me?.hasPermission(permission, {
-            checkAdmin: true,
-            checkOwner: true,
-          })
-        )
-          return new discord.MessageEmbed()
+        if (!message.guild.me?.permissions.has(permission, true))
+          return new core.SafeMessageEmbed()
             .setColor("RED")
             .setAuthor("Oops!", message.client.user.displayAvatarURL())
             .setDescription(
@@ -349,13 +370,8 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
       )
 
       for (const permission of userPermissions)
-        if (
-          !message.member.hasPermission(permission, {
-            checkAdmin: true,
-            checkOwner: true,
-          })
-        )
-          return new discord.MessageEmbed()
+        if (!message.member.permissions.has(permission, true))
+          return new core.SafeMessageEmbed()
             .setColor("RED")
             .setAuthor("Oops!", message.client.user.displayAvatarURL())
             .setDescription(
@@ -381,7 +397,7 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
           const id = getRoleId(roleCond)
 
           if (!member.roles.cache.has(id)) {
-            return new discord.MessageEmbed()
+            return new core.SafeMessageEmbed()
               .setColor("RED")
               .setAuthor("Oops!", message.client.user.displayAvatarURL())
               .setDescription(
@@ -395,7 +411,7 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
               const id = getRoleId(_roleCond)
 
               if (member.roles.cache.has(id)) {
-                return new discord.MessageEmbed()
+                return new core.SafeMessageEmbed()
                   .setColor("RED")
                   .setAuthor("Oops!", message.client.user.displayAvatarURL())
                   .setDescription(
@@ -405,7 +421,7 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
             } else {
               for (const role of _roleCond) {
                 if (member.roles.cache.has(getRoleId(role))) {
-                  return new discord.MessageEmbed()
+                  return new core.SafeMessageEmbed()
                     .setColor("RED")
                     .setAuthor("Oops!", message.client.user.displayAvatarURL())
                     .setDescription(
@@ -425,7 +441,7 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
                   `Bad command.roles structure in ${chalk.bold(
                     commandBreadcrumb(cmd, "/")
                   )} command.`,
-                  "handler"
+                  "command:prepareCommand"
                 )
               } else {
                 const id = getRoleId(role)
@@ -438,7 +454,7 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
             }
 
             if (!someRoleGiven)
-              return new discord.MessageEmbed()
+              return new core.SafeMessageEmbed()
                 .setColor("RED")
                 .setAuthor("Oops!", message.client.user.displayAvatarURL())
                 .setDescription(
@@ -460,7 +476,7 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
 
   if (channelType === "guild")
     if (isDirectMessage(message))
-      return new discord.MessageEmbed()
+      return new core.SafeMessageEmbed()
         .setColor("RED")
         .setAuthor(
           "This command must be used in a guild.",
@@ -469,7 +485,7 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
 
   if (await core.scrap(cmd.options.botOwnerOnly, message))
     if (process.env.BOT_OWNER !== message.author.id)
-      return new discord.MessageEmbed()
+      return new core.SafeMessageEmbed()
         .setColor("RED")
         .setAuthor(
           "You must be my owner.",
@@ -482,8 +498,8 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
 
       for (const positional of positionalList) {
         const index = positionalList.indexOf(positional)
-        let value = context.parsedArgs._[index]
-        const given = value !== undefined
+        let value: any = context.parsedArgs._[index]
+        const given = value !== undefined && value !== null
 
         const set = (v: any) => {
           message.args[positional.name] = v
@@ -497,7 +513,21 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
 
         if (!given) {
           if (await core.scrap(positional.required, message)) {
-            return new discord.MessageEmbed()
+            if (positional.missingErrorMessage) {
+              if (typeof positional.missingErrorMessage === "string") {
+                return new core.SafeMessageEmbed()
+                  .setColor("RED")
+                  .setAuthor(
+                    `Missing positional "${positional.name}"`,
+                    message.client.user.displayAvatarURL()
+                  )
+                  .setDescription(positional.missingErrorMessage)
+              } else {
+                return positional.missingErrorMessage
+              }
+            }
+
+            return new core.SafeMessageEmbed()
               .setColor("RED")
               .setAuthor(
                 `Missing positional "${positional.name}"`,
@@ -543,7 +573,7 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
         if (value !== null && positional.checkCastedValue) {
           const checked = await argument.checkCastedValue(
             positional,
-            "argument",
+            "positional",
             value,
             message
           )
@@ -571,11 +601,25 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
 
         if (value === true) value = undefined
 
-        if ((await core.scrap(option.required, message)) && !given)
-          return new discord.MessageEmbed()
+        if (!given && (await core.scrap(option.required, message))) {
+          if (option.missingErrorMessage) {
+            if (typeof option.missingErrorMessage === "string") {
+              return new core.SafeMessageEmbed()
+                .setColor("RED")
+                .setAuthor(
+                  `Missing option "${option.name}"`,
+                  message.client.user.displayAvatarURL()
+                )
+                .setDescription(option.missingErrorMessage)
+            } else {
+              return option.missingErrorMessage
+            }
+          }
+
+          return new core.SafeMessageEmbed()
             .setColor("RED")
             .setAuthor(
-              `Missing argument "${option.name}"`,
+              `Missing option "${option.name}"`,
               message.client.user.displayAvatarURL()
             )
             .setDescription(
@@ -583,6 +627,7 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
                 ? "Description: " + option.description
                 : `Example: \`--${option.name}=someValue\``
             )
+        }
 
         set(value)
 
@@ -630,7 +675,7 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
 
     if (cmd.options.flags) {
       for (const flag of cmd.options.flags) {
-        let { given, value } = argument.resolveGivenArgument(
+        let { given, nameIsGiven, value } = argument.resolveGivenArgument(
           context.parsedArgs,
           flag
         )
@@ -640,7 +685,7 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
           value = v
         }
 
-        if (!given) set(false)
+        if (!nameIsGiven) set(false)
         else if (typeof value === "boolean") set(value)
         else if (/^(?:true|1|on|yes|oui)$/.test(value)) set(true)
         else if (/^(?:false|0|off|no|non)$/.test(value)) set(false)
@@ -660,7 +705,21 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
 
       if (message.rest.length === 0) {
         if (await core.scrap(rest.required, message)) {
-          return new discord.MessageEmbed()
+          if (rest.missingErrorMessage) {
+            if (typeof rest.missingErrorMessage === "string") {
+              return new core.SafeMessageEmbed()
+                .setColor("RED")
+                .setAuthor(
+                  `Missing rest "${rest.name}"`,
+                  message.client.user.displayAvatarURL()
+                )
+                .setDescription(rest.missingErrorMessage)
+            } else {
+              return rest.missingErrorMessage
+            }
+          }
+
+          return new core.SafeMessageEmbed()
             .setColor("RED")
             .setAuthor(
               `Missing rest "${rest.name}"`,
@@ -682,13 +741,26 @@ export async function prepareCommand<Type extends keyof CommandMessageType>(
   if (cmd.options.middlewares) {
     const middlewares = await core.scrap(cmd.options.middlewares, message)
 
+    let currentData: any = {}
+
     for (const middleware of middlewares) {
-      const result: string | boolean = await middleware(message)
+      const { result, data } = await middleware(message, currentData)
+
+      currentData = {
+        ...currentData,
+        ...(data ?? {}),
+      }
 
       if (typeof result === "string")
-        return new discord.MessageEmbed()
+        return new core.SafeMessageEmbed()
           .setColor("RED")
-          .setAuthor(result, message.client.user.displayAvatarURL())
+          .setAuthor(
+            `${
+              middleware.name ? `"${middleware.name}" m` : "M"
+            }iddleware error`,
+            message.client.user.displayAvatarURL()
+          )
+          .setDescription(result)
 
       if (!result) return false
     }
@@ -750,10 +822,10 @@ export async function sendCommandDetails<Type extends keyof CommandMessageType>(
           : ""
       argumentList.push(
         (await core.scrap(arg.required, message))
-          ? `\`--${arg.name}${dft}\` (\`${argument.getTypeDescriptionOf(
+          ? `\`--${arg.name}${dft}\` (\`${argument.getCastingDescriptionOf(
               arg
             )}\`) ${arg.description ?? ""}`
-          : `\`[--${arg.name}${dft}]\` (\`${argument.getTypeDescriptionOf(
+          : `\`[--${arg.name}${dft}]\` (\`${argument.getCastingDescriptionOf(
               arg
             )}\`) ${arg.description ?? ""}`
       )
@@ -773,8 +845,8 @@ export async function sendCommandDetails<Type extends keyof CommandMessageType>(
   if (await core.scrap(cmd.options.guildOwnerOnly, message))
     specialPermissions.push("GUILD_OWNER")
 
-  const embed = new discord.MessageEmbed()
-    .setColor("BLURPLE")
+  const embed = new core.SafeMessageEmbed()
+    .setColor()
     .setAuthor("Command details", message.client.user.displayAvatarURL())
     .setTitle(
       `${pattern} ${[...positionalList, restPattern, ...flagList].join(" ")} ${
@@ -791,13 +863,21 @@ export async function sendCommandDetails<Type extends keyof CommandMessageType>(
     embed.addField("options", argumentList.join("\n"), false)
 
   if (cmd.options.aliases) {
-    const aliases = Array.isArray(cmd.options.aliases)
-      ? cmd.options.aliases
-      : [cmd.options.aliases]
+    const aliases = cmd.options.aliases
 
     embed.addField(
       "aliases",
       aliases.map((alias) => `\`${alias}\``).join(", "),
+      true
+    )
+  }
+
+  if (cmd.options.middlewares) {
+    embed.addField(
+      "middlewares:",
+      cmd.options.middlewares
+        .map((middleware) => `*${middleware.name || "Anonymous"}*`)
+        .join(" → "),
       true
     )
   }
@@ -866,7 +946,7 @@ export async function sendCommandDetails<Type extends keyof CommandMessageType>(
       `This command can only be sent in ${cmd.options.channelType} channel.`
     )
 
-  await message.channel.send(embed)
+  await message.channel.send({ embeds: [embed] })
 }
 
 export function commandToListItem<Type extends keyof CommandMessageType>(
@@ -878,19 +958,19 @@ export function commandToListItem<Type extends keyof CommandMessageType>(
   }`
 }
 
-export function isCommandMessage(
-  message: discord.Message
-): message is CommandMessage {
+export function isNormalMessage(
+  message: discord.Message | discord.PartialMessage
+): message is NormalMessage {
   return (
     !message.system &&
     !!message.channel &&
     !!message.author &&
-    !message.webhookID
+    !message.webhookId
   )
 }
 
 export function isGuildMessage(
-  message: CommandMessage
+  message: NormalMessage
 ): message is GuildMessage {
   return (
     !!message.member &&
@@ -900,7 +980,7 @@ export function isGuildMessage(
 }
 
 export function isDirectMessage(
-  message: CommandMessage
+  message: NormalMessage
 ): message is DirectMessage {
   return message.channel instanceof discord.DMChannel
 }
