@@ -1,41 +1,49 @@
 import discord from "discord.js"
 import path from "path"
 import chalk from "chalk"
-import apiTypes from "discord-api-types/v8"
+import apiTypes from "discord-api-types/v8.js"
 
-import * as logger from "./logger"
-import * as handler from "./handler"
+import * as logger from "./logger.js"
+import * as handler from "./handler.js"
+import * as core from "./core.js"
 
 export const listenerHandler = new handler.Handler(
   process.env.BOT_LISTENERS_PATH ??
     path.join(process.cwd(), "dist", "listeners")
 )
 
-listenerHandler.once("finish", (pathList, client) => {
-  pathList.forEach((filepath) => {
-    const listener = require(filepath)
-    client[listener.once ? "once" : "on"](
-      listener.event,
-      listener.run.bind(client)
-    )
-    logger.log(
-      `loaded listener ${chalk.yellow(
-        listener.once ? "once" : "on"
-      )} ${chalk.blueBright(listener.event)}`,
-      "handler"
-    )
+listenerHandler.on("load", async (filepath, client) => {
+  const file = await import("file://" + filepath)
+  const listener = file.default as Listener<any>
+
+  client[listener.once ? "once" : "on"](listener.event, async (...args) => {
+    try {
+      await listener.run.bind(client)(...args)
+    } catch (error: any) {
+      logger.error(error, filepath, true)
+    }
   })
+
+  const sub = path.basename(filepath, ".js").replace(`${listener.event}.`, "")
+
+  logger.log(
+    `loaded listener ${chalk.yellow(
+      listener.once ? "once" : "on"
+    )} ${chalk.blueBright(listener.event)}${
+      sub !== listener.event ? ` ${chalk.green(sub)}` : ""
+    } ${chalk.grey(listener.description)}`
+  )
 })
 
-export type Packet = apiTypes.GatewayDispatchPayload
+export interface MoreClientEvents {
+  raw: [packet: apiTypes.GatewayDispatchPayload]
+}
 
-export type Listener<
-  EventName extends keyof (discord.ClientEvents & { raw: [Packet] })
-> = {
+export type AllClientEvents = discord.ClientEvents & MoreClientEvents
+
+export type Listener<EventName extends keyof AllClientEvents> = {
   event: EventName
-  run: (
-    this: discord.Client,
-    ...args: (discord.ClientEvents & { raw: [Packet] })[EventName]
-  ) => unknown
+  description: string
+  run: (this: core.FullClient, ...args: AllClientEvents[EventName]) => unknown
   once?: boolean
 }
