@@ -1,4 +1,8 @@
-import Discord from "discord.js"
+import discord from "discord.js"
+import type { FullClient } from "./app.js"
+import { filename } from "dirname-filename-esm"
+
+const __filename = filename(import.meta)
 
 import "dotenv/config"
 
@@ -8,24 +12,29 @@ for (const key of ["BOT_TOKEN", "BOT_PREFIX", "BOT_OWNER"]) {
   }
 }
 
-const client = new Discord.Client()
+export const client = new discord.Client({
+  intents: process.env.BOT_INTENTS
+    ? process.env.BOT_INTENTS.split(/[;|.,\s+]+/).map(
+        (intent) => discord.Intents.FLAGS[intent as discord.IntentsString]
+      )
+    : [],
+})
 
-;(async () => {
-  const app = await import("./app")
+const app = await import("./app.js")
 
-  try {
-    await client.login(process.env.BOT_TOKEN)
+try {
+  await app.tableHandler.load(client as FullClient)
+  await app.commandHandler.load(client as FullClient)
+  await app.listenerHandler.load(client as FullClient)
 
-    if (!app.isFullClient(client)) {
-      app.error("The Discord client is not full.", "system")
-      client.destroy()
-      process.exit(1)
-    }
+  await client.login(process.env.BOT_TOKEN)
 
-    await app.tableHandler.load(client)
-    await app.commandHandler.load(client)
-    await app.listenerHandler.load(client)
-  } catch (error) {
-    app.error(error, "system", true)
+  if (!app.isFullClient(client)) {
+    app.error("The Discord client is not full.", __filename)
+    client.destroy()
+    process.exit(1)
   }
-})()
+} catch (error: any) {
+  app.error(error, __filename, true)
+  process.exit(1)
+}
